Update doc example with combine(groupby(...
Replaced `data_plot =DataFrames.by(...` with `data_plot =combine(groupby(...` in the doc example.

To note that the file updated do not correspond to the built documentation. In the documentation currently displayed on the site there are other errors that have already been corrected in the file I am editing now... check the output of your github actions!